AdvancedAgents/SocialMediaVideoEdit-Gem/backend/requirements.txt
fastapi
uvicorn
python-multipart
ffmpeg-python
google-generativeai
python-dotenv
websockets
